requirements.txt
boto3
blinker
certifi
charset-normalizer
click
colorama
Flask
idna
itsdangerous
Jinja2
MarkupSafe
requests
urllib3
Werkzeug
psycopg2-binary
python-dotenv
